fix: create default config.toml instead of error

+++ cmd/start.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"errors"
 	"os"
+	"path/filepath"
 	"time"
 
 	"github.com/getsentry/sentry-go"
@@ -27,17 +28,26 @@ var startCmd = &cobra.Command{
 
 		file, err := cmd.Flags().GetString("config")
 		if err != nil {
-			log.Info("unable to use given config file:", err)
-			log.Info("Creating default config.toml")
-			err = os.WriteFile("config.toml", []byte{}, 0644)
-			if err != nil {
-				return err
-			}
+			return err
 		}
 		if file == "" {
 			return errors.New("you must specify a config file with '--config PATH'")
 		}
+		if file == "config.toml" {
+			filename, err := filepath.Abs(file)
+			if err != nil {
+				return err
+			}
 
+			if _, err := os.Stat(filename); errors.Is(err, os.ErrNotExist) {
+				log.Info("unable to use given config file:", err)
+				log.Info("Creating default config.toml")
+				err = os.WriteFile("config.toml", []byte{}, 0644)
+				if err != nil {
+					return err
+				}
+			}
+		}
 		regions, err := cmd.Flags().GetStringArray("regions")
 		if err != nil {
 			return err
